+++ framework/contracts/native/ans-host/src/commands.rs
@@ -363,6 +363,7 @@ fn validate_pool_assets(
 
 #[cfg(test)]
 mod test {
+    #![allow(clippy::needless_borrows_for_generic_args)]
     use abstract_testing::{map_tester::CwMapTester, prelude::*};
     use cosmwasm_std::{
         testing::{mock_dependencies, mock_env, mock_info},