+++ framework/packages/abstract-adapter/src/endpoints/reply.rs
@@ -9,7 +9,8 @@ impl<Error: ContractError, CustomInitMsg, CustomExecMsg, CustomQueryMsg, Receive
 }
 
 #[cfg(test)]
-mod tests {
+mod test {
+    #![allow(clippy::needless_borrows_for_generic_args)]
     use abstract_sdk::AbstractSdkError;
     use cosmwasm_std::{
         testing::{mock_dependencies, mock_env},