+++ framework/packages/abstract-std/src/objects/namespace.rs
@@ -116,6 +116,7 @@ impl KeyDeserialize for Namespace {
 
 #[cfg(test)]
 mod test {
+    #![allow(clippy::needless_borrows_for_generic_args)]
     use speculoos::prelude::*;
 
     use super::*;