+++ framework/packages/abstract-std/src/objects/pool/pool_id.rs
@@ -176,6 +176,7 @@ impl fmt::Display for PoolAddress {
 
 #[cfg(test)]
 mod test {
+    #![allow(clippy::needless_borrows_for_generic_args)]
     use cosmwasm_std::testing::MockApi;
     use speculoos::prelude::*;