+++ modules/contracts/adapters/dex/src/api.rs
@@ -1,14 +1,16 @@
 use crate::DEX_ADAPTER_ID;
+use abstract_adapter::sdk::Resolve;
 use abstract_adapter::sdk::{
     features::{AccountIdentification, Dependencies, ModuleIdentification},
     AbstractSdkResult, AdapterInterface,
 };
 use abstract_adapter::std::objects::{module::ModuleId, AnsAsset, AssetEntry, PoolAddress};
+
+use abstract_adapter::traits::AbstractNameService;
 use abstract_dex_standard::msg::GenerateMessagesResponse;
 use abstract_dex_standard::{
-    ans_action::DexAnsAction,
+    action::DexAction,
     msg::{DexExecuteMsg, DexName, DexQueryMsg, SimulateSwapResponse},
-    raw_action::DexRawAction,
 };
 use cosmwasm_schema::serde::de::DeserializeOwned;
 use cosmwasm_std::{CosmosMsg, Decimal, Deps};
@@ -18,7 +20,9 @@ use self::{ans::AnsDex, raw::Dex};
 
 // API for Abstract SDK users
 /// Interact with the dex adapter in your module.
-pub trait DexInterface: AccountIdentification + Dependencies + ModuleIdentification {
+pub trait DexInterface:
+    AccountIdentification + Dependencies + ModuleIdentification + AbstractNameService
+{
     /// Construct a new dex interface.
     fn dex<'a>(&'a self, deps: Deps<'a>, name: DexName) -> Dex<Self> {
         Dex {
@@ -39,7 +43,10 @@ pub trait DexInterface: AccountIdentification + Dependencies + ModuleIdentificat
     }
 }
 
-impl<T: AccountIdentification + Dependencies + ModuleIdentification> DexInterface for T {}
+impl<T: AccountIdentification + Dependencies + ModuleIdentification + AbstractNameService>
+    DexInterface for T
+{
+}
 
 pub mod raw {
     use super::*;
@@ -78,13 +85,13 @@ pub mod raw {
             self.module_id
         }
 
-        /// Executes a [DexRawAction] in th DEX
-        fn execute(&self, action: DexRawAction) -> AbstractSdkResult<CosmosMsg> {
+        /// Executes a [DexAction] in th DEX
+        fn execute(&self, action: DexAction) -> AbstractSdkResult<CosmosMsg> {
             let adapters = self.base.adapters(self.deps);
 
             adapters.execute(
                 self.dex_module_id(),
-                DexExecuteMsg::RawAction {
+                DexExecuteMsg::Action {
                     dex: self.dex_name(),
                     action,
                 },
@@ -100,7 +107,7 @@ pub mod raw {
             belief_price: Option<Decimal>,
             pool: PoolAddress,
         ) -> AbstractSdkResult<CosmosMsg> {
-            self.execute(DexRawAction::Swap {
+            self.execute(DexAction::Swap {
                 offer_asset: offer_asset.into(),
                 ask_asset: ask_asset.into(),
                 belief_price,
@@ -116,24 +123,10 @@ pub mod raw {
             max_spread: Option<Decimal>,
             pool: PoolAddress,
         ) -> AbstractSdkResult<CosmosMsg> {
-            self.execute(DexRawAction::ProvideLiquidity {
+            self.execute(DexAction::ProvideLiquidity {
                 assets: assets.into_iter().map(Into::into).collect(),
                 pool: pool.into(),
                 max_spread,
-            })
-        }
-
-        /// Provide symmetric liquidity in the DEX
-        pub fn provide_liquidity_symmetric(
-            &self,
-            offer_asset: Asset,
-            paired_assets: Vec<AssetInfo>,
-            pool: PoolAddress,
-        ) -> AbstractSdkResult<CosmosMsg> {
-            self.execute(DexRawAction::ProvideLiquiditySymmetric {
-                offer_asset: offer_asset.into(),
-                paired_assets: paired_assets.into_iter().map(Into::into).collect(),
-                pool: pool.into(),
             })
         }
 
@@ -143,7 +136,7 @@ pub mod raw {
             lp_token: Asset,
             pool: PoolAddress,
         ) -> AbstractSdkResult<CosmosMsg> {
-            self.execute(DexRawAction::WithdrawLiquidity {
+            self.execute(DexAction::WithdrawLiquidity {
                 lp_token: lp_token.into(),
                 pool: pool.into(),
             })
@@ -185,9 +178,9 @@ pub mod raw {
             addr_as_sender: impl Into<String>,
         ) -> AbstractSdkResult<GenerateMessagesResponse> {
             let response: GenerateMessagesResponse = self.query(DexQueryMsg::GenerateMessages {
-                message: DexExecuteMsg::RawAction {
+                message: DexExecuteMsg::Action {
                     dex: self.dex_name(),
-                    action: DexRawAction::Swap {
+                    action: DexAction::Swap {
                         offer_asset: offer_asset.into(),
                         ask_asset: ask_asset.into(),
                         max_spread,
@@ -203,6 +196,9 @@ pub mod raw {
 }
 
 pub mod ans {
+    use abstract_adapter::std::AbstractError;
+    use abstract_dex_standard::ans_action::{DexAnsAction, WholeDexAction};
+
     use super::*;
 
     #[derive(Clone)]
@@ -242,14 +238,12 @@ pub mod ans {
         /// Executes a [DexAction] in th DEX
         fn execute(&self, action: DexAnsAction) -> AbstractSdkResult<CosmosMsg> {
             let adapters = self.base.adapters(self.deps);
+            let ans_host = self.base.ans_host(self.deps)?;
+            let message = WholeDexAction(self.dex_name(), action)
+                .resolve(&self.deps.querier, &ans_host)
+                .map_err(AbstractError::from)?;
 
-            adapters.execute(
-                self.dex_module_id(),
-                DexExecuteMsg::AnsAction {
-                    dex: self.dex_name(),
-                    action,
-                },
-            )
+            adapters.execute(self.dex_module_id(), message)
         }
 
         /// Swap assets in the DEX
@@ -275,18 +269,6 @@ pub mod ans {
             max_spread: Option<Decimal>,
         ) -> AbstractSdkResult<CosmosMsg> {
             self.execute(DexAnsAction::ProvideLiquidity { assets, max_spread })
-        }
-
-        /// Provide symmetrict liquidity in the DEX
-        pub fn provide_liquidity_symmetric(
-            &self,
-            offer_asset: AnsAsset,
-            paired_assets: Vec<AssetEntry>,
-        ) -> AbstractSdkResult<CosmosMsg> {
-            self.execute(DexAnsAction::ProvideLiquiditySymmetric {
-                offer_asset,
-                paired_assets,
-            })
         }
 
         /// Withdraw liquidity from the DEX
@@ -325,16 +307,21 @@ pub mod ans {
             belief_price: Option<Decimal>,
             addr_as_sender: impl Into<String>,
         ) -> AbstractSdkResult<GenerateMessagesResponse> {
-            let response: GenerateMessagesResponse = self.query(DexQueryMsg::GenerateMessages {
-                message: DexExecuteMsg::AnsAction {
-                    dex: self.dex_name(),
-                    action: DexAnsAction::Swap {
-                        offer_asset,
-                        ask_asset,
-                        max_spread,
-                        belief_price,
-                    },
+            let ans_host = self.base.ans_host(self.deps)?;
+            let message = WholeDexAction(
+                self.dex_name(),
+                DexAnsAction::Swap {
+                    offer_asset,
+                    ask_asset,
+                    max_spread,
+                    belief_price,
                 },
+            )
+            .resolve(&self.deps.querier, &ans_host)
+            .map_err(AbstractError::from)?;
+
+            let response: GenerateMessagesResponse = self.query(DexQueryMsg::GenerateMessages {
+                message,
                 addr_as_sender: addr_as_sender.into(),
             })?;
             Ok(response)
@@ -344,13 +331,16 @@ pub mod ans {
 
 #[cfg(test)]
 mod test {
+    #![allow(clippy::needless_borrows_for_generic_args)]
+    use super::*;
+    use crate::msg::ExecuteMsg;
     use abstract_adapter::sdk::mock_module::MockModule;
     use abstract_adapter::std::adapter::AdapterRequestMsg;
+    use abstract_adapter::std::objects::pool_id::PoolAddressBase;
     use cosmwasm_std::{testing::mock_dependencies, wasm_execute};
     use speculoos::prelude::*;
 
-    use super::*;
-    use crate::msg::ExecuteMsg;
+    pub const POOL: u64 = 1278734;
 
     fn expected_request_with_test_proxy(request: DexExecuteMsg) -> ExecuteMsg {
         AdapterRequestMsg {
@@ -362,32 +352,36 @@ mod test {
         .into()
     }
 
+    // Only raw actions tested here, since it's pointless to resolve ans query in Mock tests
+
     #[test]
     fn swap_msg() {
         let mut deps = mock_dependencies();
         deps.querier = abstract_adapter::abstract_testing::mock_querier();
         let stub = MockModule::new();
         let dex = stub
-            .ans_dex(deps.as_ref(), "junoswap".into())
+            .dex(deps.as_ref(), "junoswap".into())
             .with_module_id(abstract_adapter::abstract_testing::prelude::TEST_MODULE_ID);
 
         let dex_name = "junoswap".to_string();
-        let offer_asset = AnsAsset::new("juno", 1000u128);
-        let ask_asset = AssetEntry::new("uusd");
+        let offer_asset = Asset::native("ujuno", 100_000u128);
+        let ask_asset = AssetInfo::native("uusd");
         let max_spread = Some(Decimal::percent(1));
         let belief_price = Some(Decimal::percent(2));
+        let pool = PoolAddressBase::Id(POOL);
 
-        let expected = expected_request_with_test_proxy(DexExecuteMsg::AnsAction {
+        let expected = expected_request_with_test_proxy(DexExecuteMsg::Action {
             dex: dex_name,
-            action: DexAnsAction::Swap {
-                offer_asset: offer_asset.clone(),
-                ask_asset: ask_asset.clone(),
+            action: DexAction::Swap {
+                offer_asset: offer_asset.clone().into(),
+                ask_asset: ask_asset.clone().into(),
                 max_spread,
                 belief_price,
+                pool: pool.clone().into(),
             },
         });
 
-        let actual = dex.swap(offer_asset, ask_asset, max_spread, belief_price);
+        let actual = dex.swap(offer_asset, ask_asset, max_spread, belief_price, pool);
 
         assert_that!(actual).is_ok();
 
@@ -413,62 +407,23 @@ mod test {
         let dex_name = "junoswap".to_string();
 
         let dex = stub
-            .ans_dex(deps.as_ref(), dex_name.clone())
+            .dex(deps.as_ref(), dex_name.clone())
             .with_module_id(abstract_adapter::abstract_testing::prelude::TEST_MODULE_ID);
 
-        let assets = vec![AnsAsset::new("taco", 1000u128)];
+        let assets = vec![Asset::native("taco", 1000u128)];
         let max_spread = Some(Decimal::percent(1));
+        let pool = PoolAddressBase::Id(POOL);
 
-        let expected = expected_request_with_test_proxy(DexExecuteMsg::AnsAction {
+        let expected = expected_request_with_test_proxy(DexExecuteMsg::Action {
             dex: dex_name,
-            action: DexAnsAction::ProvideLiquidity {
-                assets: assets.clone(),
+            action: DexAction::ProvideLiquidity {
+                assets: assets.clone().into_iter().map(Into::into).collect(),
                 max_spread,
+                pool: pool.clone().into(),
             },
         });
 
-        let actual = dex.provide_liquidity(assets, max_spread);
-
-        assert_that!(actual).is_ok();
-
-        let actual = match actual.unwrap() {
-            CosmosMsg::Wasm(msg) => msg,
-            _ => panic!("expected wasm msg"),
-        };
-        let expected = wasm_execute(
-            abstract_adapter::abstract_testing::prelude::TEST_MODULE_ADDRESS,
-            &expected,
-            vec![],
-        )
-        .unwrap();
-
-        assert_that!(actual).is_equal_to(expected);
-    }
-
-    #[test]
-    fn provide_liquidity_symmetric_msg() {
-        let mut deps = mock_dependencies();
-        deps.querier = abstract_adapter::abstract_testing::mock_querier();
-        let stub = MockModule::new();
-        let dex_name = "junoswap".to_string();
-
-        let dex = stub
-            .ans_dex(deps.as_ref(), dex_name.clone())
-            .with_module_id(abstract_adapter::abstract_testing::prelude::TEST_MODULE_ID);
-
-        let offer = AnsAsset::new("taco", 1000u128);
-        let paired = vec![AssetEntry::new("bell")];
-        let _max_spread = Some(Decimal::percent(1));
-
-        let expected = expected_request_with_test_proxy(DexExecuteMsg::AnsAction {
-            dex: dex_name,
-            action: DexAnsAction::ProvideLiquiditySymmetric {
-                offer_asset: offer.clone(),
-                paired_assets: paired.clone(),
-            },
-        });
-
-        let actual = dex.provide_liquidity_symmetric(offer, paired);
+        let actual = dex.provide_liquidity(assets, max_spread, pool);
 
         assert_that!(actual).is_ok();
 
@@ -494,19 +449,21 @@ mod test {
         let dex_name = "junoswap".to_string();
 
         let dex = stub
-            .ans_dex(deps.as_ref(), dex_name.clone())
+            .dex(deps.as_ref(), dex_name.clone())
             .with_module_id(abstract_adapter::abstract_testing::prelude::TEST_MODULE_ID);
 
-        let lp_token = AnsAsset::new("taco", 1000u128);
+        let lp_token = Asset::native("taco", 1000u128);
+        let pool = PoolAddressBase::Id(POOL);
 
-        let expected = expected_request_with_test_proxy(DexExecuteMsg::AnsAction {
+        let expected = expected_request_with_test_proxy(DexExecuteMsg::Action {
             dex: dex_name,
-            action: DexAnsAction::WithdrawLiquidity {
-                lp_token: lp_token.clone(),
+            action: DexAction::WithdrawLiquidity {
+                lp_token: lp_token.clone().into(),
+                pool: pool.clone().into(),
             },
         });
 
-        let actual = dex.withdraw_liquidity(lp_token);
+        let actual = dex.withdraw_liquidity(lp_token, pool);
 
         assert_that!(actual).is_ok();
 
@@ -522,183 +479,5 @@ mod test {
         .unwrap();
 
         assert_that!(actual).is_equal_to(expected);
-    }
-
-    mod raw {
-        use abstract_adapter::std::objects::pool_id::PoolAddressBase;
-
-        use super::*;
-
-        pub const POOL: u64 = 1278734;
-
-        #[test]
-        fn swap_msg() {
-            let mut deps = mock_dependencies();
-            deps.querier = abstract_adapter::abstract_testing::mock_querier();
-            let stub = MockModule::new();
-            let dex = stub
-                .dex(deps.as_ref(), "junoswap".into())
-                .with_module_id(abstract_adapter::abstract_testing::prelude::TEST_MODULE_ID);
-
-            let dex_name = "junoswap".to_string();
-            let offer_asset = Asset::native("ujuno", 100_000u128);
-            let ask_asset = AssetInfo::native("uusd");
-            let max_spread = Some(Decimal::percent(1));
-            let belief_price = Some(Decimal::percent(2));
-            let pool = PoolAddressBase::Id(POOL);
-
-            let expected = expected_request_with_test_proxy(DexExecuteMsg::RawAction {
-                dex: dex_name,
-                action: DexRawAction::Swap {
-                    offer_asset: offer_asset.clone().into(),
-                    ask_asset: ask_asset.clone().into(),
-                    max_spread,
-                    belief_price,
-                    pool: pool.clone().into(),
-                },
-            });
-
-            let actual = dex.swap(offer_asset, ask_asset, max_spread, belief_price, pool);
-
-            assert_that!(actual).is_ok();
-
-            let actual = match actual.unwrap() {
-                CosmosMsg::Wasm(msg) => msg,
-                _ => panic!("expected wasm msg"),
-            };
-            let expected = wasm_execute(
-                abstract_adapter::abstract_testing::prelude::TEST_MODULE_ADDRESS,
-                &expected,
-                vec![],
-            )
-            .unwrap();
-
-            assert_that!(actual).is_equal_to(expected);
-        }
-
-        #[test]
-        fn provide_liquidity_msg() {
-            let mut deps = mock_dependencies();
-            deps.querier = abstract_adapter::abstract_testing::mock_querier();
-            let stub = MockModule::new();
-            let dex_name = "junoswap".to_string();
-
-            let dex = stub
-                .dex(deps.as_ref(), dex_name.clone())
-                .with_module_id(abstract_adapter::abstract_testing::prelude::TEST_MODULE_ID);
-
-            let assets = vec![Asset::native("taco", 1000u128)];
-            let max_spread = Some(Decimal::percent(1));
-            let pool = PoolAddressBase::Id(POOL);
-
-            let expected = expected_request_with_test_proxy(DexExecuteMsg::RawAction {
-                dex: dex_name,
-                action: DexRawAction::ProvideLiquidity {
-                    assets: assets.clone().into_iter().map(Into::into).collect(),
-                    max_spread,
-                    pool: pool.clone().into(),
-                },
-            });
-
-            let actual = dex.provide_liquidity(assets, max_spread, pool);
-
-            assert_that!(actual).is_ok();
-
-            let actual = match actual.unwrap() {
-                CosmosMsg::Wasm(msg) => msg,
-                _ => panic!("expected wasm msg"),
-            };
-            let expected = wasm_execute(
-                abstract_adapter::abstract_testing::prelude::TEST_MODULE_ADDRESS,
-                &expected,
-                vec![],
-            )
-            .unwrap();
-
-            assert_that!(actual).is_equal_to(expected);
-        }
-
-        #[test]
-        fn provide_liquidity_symmetric_msg() {
-            let mut deps = mock_dependencies();
-            deps.querier = abstract_adapter::abstract_testing::mock_querier();
-            let stub = MockModule::new();
-            let dex_name = "junoswap".to_string();
-
-            let dex = stub
-                .dex(deps.as_ref(), dex_name.clone())
-                .with_module_id(abstract_adapter::abstract_testing::prelude::TEST_MODULE_ID);
-
-            let offer = Asset::native("taco", 1000u128);
-            let paired = vec![AssetInfo::native("bell")];
-            let _max_spread = Some(Decimal::percent(1));
-            let pool = PoolAddressBase::Id(POOL);
-
-            let expected = expected_request_with_test_proxy(DexExecuteMsg::RawAction {
-                dex: dex_name,
-                action: DexRawAction::ProvideLiquiditySymmetric {
-                    offer_asset: offer.clone().into(),
-                    paired_assets: paired.clone().into_iter().map(Into::into).collect(),
-                    pool: pool.clone().into(),
-                },
-            });
-
-            let actual = dex.provide_liquidity_symmetric(offer, paired, pool);
-
-            assert_that!(actual).is_ok();
-
-            let actual = match actual.unwrap() {
-                CosmosMsg::Wasm(msg) => msg,
-                _ => panic!("expected wasm msg"),
-            };
-            let expected = wasm_execute(
-                abstract_adapter::abstract_testing::prelude::TEST_MODULE_ADDRESS,
-                &expected,
-                vec![],
-            )
-            .unwrap();
-
-            assert_that!(actual).is_equal_to(expected);
-        }
-
-        #[test]
-        fn withdraw_liquidity_msg() {
-            let mut deps = mock_dependencies();
-            deps.querier = abstract_adapter::abstract_testing::mock_querier();
-            let stub = MockModule::new();
-            let dex_name = "junoswap".to_string();
-
-            let dex = stub
-                .dex(deps.as_ref(), dex_name.clone())
-                .with_module_id(abstract_adapter::abstract_testing::prelude::TEST_MODULE_ID);
-
-            let lp_token = Asset::native("taco", 1000u128);
-            let pool = PoolAddressBase::Id(POOL);
-
-            let expected = expected_request_with_test_proxy(DexExecuteMsg::RawAction {
-                dex: dex_name,
-                action: DexRawAction::WithdrawLiquidity {
-                    lp_token: lp_token.clone().into(),
-                    pool: pool.clone().into(),
-                },
-            });
-
-            let actual = dex.withdraw_liquidity(lp_token, pool);
-
-            assert_that!(actual).is_ok();
-
-            let actual = match actual.unwrap() {
-                CosmosMsg::Wasm(msg) => msg,
-                _ => panic!("expected wasm msg"),
-            };
-            let expected = wasm_execute(
-                abstract_adapter::abstract_testing::prelude::TEST_MODULE_ADDRESS,
-                &expected,
-                vec![],
-            )
-            .unwrap();
-
-            assert_that!(actual).is_equal_to(expected);
-        }
     }
 }