+++ modules/contracts/adapters/dex/src/exchanges/mod.rs
@@ -1,3 +1,1 @@
-pub mod junoswap;
-
 pub(crate) mod exchange_resolver;