+++ scripts/wasm-all.sh
@@ -2,13 +2,9 @@
 
 # Detect the architecture #
 if [[ $(arch) == "arm64" ]]; then
-image="cosmwasm/rust-optimizer-arm64"
-workspace_image="cosmwasm/workspace-optimizer-arm64"
-abstract_image="abstractmoney/workspace-optimizer-arm64"
+  image="cosmwasm/optimizer-arm64:0.16.0"
 else
-image="cosmwasm/rust-optimizer"
-workspace_image="cosmwasm/workspace-optimizer"
-abstract_image="abstractmoney/workspace-optimizer"
+  image="cosmwasm/optimizer:0.16.0"
 fi
 
 starting_dir=$(pwd)
@@ -20,9 +16,9 @@ cd ./framework
 rm -rf ./artifacts/*.wasm
 # Optimized builds
 docker run --rm -v "$(pwd)":/code \
---mount type=volume,source="$(basename "$(pwd)")_cache",target=/code/target \
---mount type=volume,source=registry_cache,target=/usr/local/cargo/registry \
-${workspace_image}:0.15.0
+  --mount type=volume,source="$(basename "$(pwd)")_cache",target=/code/target \
+  --mount type=volume,source=registry_cache,target=/usr/local/cargo/registry \
+  ${image}
 
 cd $starting_dir
 
@@ -37,4 +33,4 @@ docker run --rm -v "$(pwd)":/code \
   -v "$(dirname "$(pwd)")/integrations":/integrations \
   -v "$(dirname "$(pwd)")/framework":/framework \
   --mount type=volume,source=registry_cache,target=/usr/local/cargo/registry \
-  ${abstract_image}:0.15.0
+  ${image}