+++ modules/CHANGELOG.md
@@ -11,7 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- Dex Raw Action renamed to Dex Action
+
 ### Removed
+
+- Dex Ans Action from Dex Adapters, now resolve should be done by the caller of the Dex Adapter
+- Provide liquidity symmetric no longer supported on Dex Adapters
 
 ### Fixed